requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Validation and Settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utils
python-dotenv==1.0.0
openpyxl==3.1.2
pandas==2.1.4

# Web scraping (original dependencies)
requests==2.31.0
lxml==4.9.3
zeep==4.2.1